+++ src/server/components/feedback/__tests__/controllers.test.ts
@@ -1,4 +1,4 @@
-import { feedbackIngest } from "../feedback";
+import { feedbackIngest } from "../controllers";
 import * as feedbackModel from "server/models/feedback";
 
 describe("Lists Controllers", () => {
