+++ src/server/components/lists/searches/covid-test-provider.ts
@@ -1,13 +1,13 @@
 import { Request, Response } from "express";
 import { listItem } from "server/models";
-import { DEFAULT_VIEW_PROPS } from "./constants";
+import { DEFAULT_VIEW_PROPS } from "../constants";
 import {
   getServiceLabel,
   getAllRequestParams,
   removeQueryParameter,
   queryStringFromParams,
-} from "./helpers";
-import { QuestionName } from "./types";
+} from "../helpers";
+import { QuestionName } from "../types";
 
 export const covidTestProviderQuestionsSequence = [
   QuestionName.readNotice,
